+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Type parameter for `onError` added, so a custom error type can be used without changing [App.Error](https://kit.svelte.dev/docs/errors#type-safety).
+
 ## [2.19.0]
 
 ### Changed